requirements.txt
openai>=1.0.0
beautifulsoup4
requests==2.31.0
markdown==3.4.3
pandas==2.1.1
sqlite3-wrapper
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs
frozenlist==1.4.0
multidict==6.0.4
yarl==1.9.3
pydantic>=2.0.0
python-dotenv>=1.0.0
python-docx==0.8.11
Jinja2==3.1.2
html2docx==1.6.0